ts-gcp-serverless-pubsub-bigtable/application-code/frontend-app/requirements.txt
google-cloud-pubsub>=2.15.1